tugas-cassandra/requirements.txt
flask == 1.0.2
cassandra-driver